consolidate gzip header constants into encoder.go and helpers into decoder.go
Drops kafka_gzip-header.go: GzipHeaderKey / GzipHeaderValue move next to
NewGzipEncoder (producer-side), GzipActive / RemoveGzipHeader move next
to GzipDecoder (consumer-side). Tests for the helpers move into
kafka_gzip-decoder_test.go. Public API unchanged.
